requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-community==0.0.13
langchain-experimental==0.0.49
ollama==0.1.7
pydantic==2.5.2
requests==2.31.0
python-dotenv==1.0.0
httpx==0.26.0
tavily-python==0.3.3